Add flags support to data filtering
This allows filtering flag-like columns using hexadecimal filter strings
to reduce the result set to rows matching all specified flags. For
example, filtering the Achievement tables' Flags field with the string
"0x1004" will only return rows that have those bits set in the field.

In its current implementation flags matching is exhaustive; a filter of
"0x1004" requires that both the 0x1000 and 0x4 bits be set for a row to
pass the filter. This is probably the more desirable default logic.

If a user attempts to filter the flags of a non-numeric field then an
exception will bubble up to the request handler and no rows will be
returned.